requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
dataclasses-json==0.6.4
docutils==0.20.1
et-xmlfile==1.1.0
Flask==3.0.2
fonttools==4.48.1
frozenlist==1.4.1
greenlet==3.0.3
ibm-cloud-sdk-core==3.19.1
ibm-cos-sdk==2.13.4
ibm-cos-sdk-core==2.13.4
ibm-cos-sdk-s3transfer==2.13.4
ibm-db==3.2.1
ibm-watson==7.0.1
ibm-watson-machine-learning==1.0.345
idna==3.6
fastapi
uvicorn
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
kiwisolver==1.4.5
langchain==0.1.6
langchain-community==0.0.19
langchain-core==0.1.22
langsmith==0.0.87
ldap3==2.9.1
lockfile==0.12.2
lomond==0.3.3
MarkupSafe==2.1.5
marshmallow==3.20.2
matplotlib==3.8.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openpyxl==3.1.2
packaging==23.2
pandas==1.5.3
pillow==10.2.0
plotly==5.18.0
pyasn1==0.5.1
pydantic==2.6.1
pydantic_core==2.16.2
PyJWT==2.8.0
pyparsing==3.1.1
python-daemon==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
scikit-learn==1.4.0
scipy==1.12.0
six==1.16.0
slackclient==2.9.4
sniffio==1.3.0
SQLAlchemy==2.0.26
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.2.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.1.0
websocket-client==1.7.0
Werkzeug==3.0.1
yarl==1.9.4
zipp==3.17.0
